backend/requirements.txt
numpy
scikit-fuzzy
paho-mqtt